infra/exchanges/bybit/candles_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pandas==2.2.2
python-dateutil==2.9.0.post0
pytest==8.3.2
httpx==0.27.2
